requirements.txt
aiogram==3.13.1
aiohttp==3.10.2
sqlalchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.1
apscheduler==3.10.4
yookassa==3.2.0
redis==5.0.8
uvloop==0.20.0
loguru==0.7.2
psycopg2-binary==2.9.9 
Pillow==10.4.0
qrcode==8.0
python-multipart==0.0.12
pydantic==2.9.2
psutil==6.1.0
fastapi==0.115.2
uvicorn==0.32.0 